src/main/java/com/gandalf/EmailTemplate.java
package com.gandalf;

public class EmailTemplate {
    private String id;
    private String title;
    private String subject;
    private String body;
    private String accountId;


    public static EmailTemplate Default(){
        return new EmailTemplate();
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getSubject() {
        return subject;
    }

    public void setSubject(String subject) {
        this.subject = subject;
    }

    public String getAccountId() {
        return accountId;
    }

    public void setAccountId(String accountId) {
        this.accountId = accountId;
    }

    public String getBody() {
        return body;
    }

    public void setBody(String body) {
        this.body = body;
    }
}